+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Refactor circle and bubble views, now ChartPointEllipseView
+- Rename BubbleView in InfoBubble and move it to example project
 
 ## [0.2.1] - 2015-05-24
 - Improve performance drawing guidelines